# tmi-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TMI (Threat Modeling Improved) API",
    "description": "A RESTful API for collaborative threat modeling with full X6 graph library compatibility. This API provides schemas that align with AntV X6 cell object models for seamless integration with modern diagramming libraries. Supports OAuth 2.0 authentication with client callback integration for seamless single-page application authentication flows.",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "https://api.tmi.example.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authenticated requests"
      },
      "websocketAuth": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "JWT token for WebSocket authentication. WebSocket connections must include the JWT token as a query parameter since browsers do not support custom headers for WebSocket connections."
      }
    },
    "schemas": {
...
      "PaginatedResponse": {
        "type": "object",
        "description": "Paginated response wrapper for list endpoints",
        "properties": {
          "data": {
            "type": "array",
            "description": "Array of items for the current page"
          },
          "pagination": {
            "type": "object",
            "description": "Pagination metadata",
            "properties": {
              "total": {
                "type": "integer",
                "description": "Total number of items across all pages",
                "minimum": 0
              },
              "limit": {
                "type": "integer",
                "description": "Maximum number of items per page",
                "minimum": 1
              },
              "offset": {
                "type": "integer",
                "description": "Number of items skipped from the beginning",
                "minimum": 0
              },
              "has_more": {
                "type": "boolean",
                "description": "Whether there are more items available beyond this page"
              }
            },
            "required": ["total", "limit", "offset", "has_more"]
          }
        },
        "required": ["data", "pagination"]
      },
...
    },
...
  },
  "paths": {
...
    "/threat_models": {
      "get": {
        "tags": ["Threat Models"],
        "summary": "List threat models",
        "description": "Returns a list of threat models accessible to the authenticated user with enhanced metadata and entity counts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Number of threat models to return"
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Pagination offset"
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "created_at:desc"
            },
            "description": "Sort order (e.g., created_at:desc, name:asc)"
          },
          {
            "name": "owner",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by owner username or email"
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by threat model name (partial match)"
          },
          {
            "name": "description",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by threat model description (partial match)"
          },
          {
            "name": "issue_url",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by issue URL (partial match)"
          },
          {
            "name": "created_after",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Filter threat models created after this date (RFC3339 format)"
          },
          {
            "name": "created_before",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Filter threat models created before this date (RFC3339 format)"
          },
          {
            "name": "modified_after",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Filter threat models modified after this date (RFC3339 format)"
          },
          {
            "name": "modified_before",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Filter threat models modified before this date (RFC3339 format)"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of threat models with enhanced metadata and entity counts",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TMListItem"
                          }
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "data": [
                    {
                      "id": "550e8400-e29b-41d4-a716-446655440000",
                      "name": "System Threat Model",
                      "description": "Comprehensive threat model for our web application system",
                      "created_at": "2025-04-06T12:00:00Z",
                      "modified_at": "2025-04-06T14:30:00Z",
                      "owner": "user@example.com",
                      "created_by": "admin@example.com",
                      "threat_model_framework": "STRIDE",
                      "issue_url": "https://github.com/example/issues/123",
                      "document_count": 3,
                      "source_count": 2,
                      "diagram_count": 1,
                      "threat_count": 15
                    }
                  ],
                  "pagination": {
                    "total": 1,
                    "limit": 20,
                    "offset": 0,
                    "has_more": false
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "tags": ["Threat Models"],
        "summary": "Create a threat model",
        "description": "Creates a new threat model with the authenticated user as owner",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the threat model"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the threat model"
                  }
                },
                "required": ["name"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Threat model created successfully",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "URL of the new threat model"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreatModel"
                },
                "example": {
                  "id": "550e8400-e29b-41d4-a716-446655440000",
                  "name": "System Threat Model",
                  "description": "Threats for system X",
                  "created_at": "2025-04-06T12:00:00Z",
                  "modified_at": "2025-04-06T12:00:00Z",
                  "owner": "user@example.com",
                  "created_by": "user@example.com",
                  "threat_model_framework": "STRIDE",
                  "authorization": [
                    {
                      "subject": "user@example.com",
                      "role": "owner"
                    }
                  ],
                  "metadata": [],
                  "documents": [],
                  "sourceCode": [],
                  "diagrams": [],
                  "threats": []
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/threat_models/{threat_model_id}/diagrams": {
      "get": {
        "tags": ["Threat Model Sub-Resources"],
        "summary": "List threat model diagrams",
        "description": "Returns all diagrams associated with a specific threat model",
        "operationId": "getThreatModelDiagrams",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "threat_model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "maxLength": 36
            },
            "description": "Unique identifier of the threat model (UUID)"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Maximum number of items to return"
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "List of diagrams",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiagramListItem"
                  }
                },
                "example": [
                  {
                    "id": "123e4567-e89b-12d3-a456-426614174000",
                    "name": "System Architecture Diagram",
                    "type": "DFD-1.0.0"
                  },
                  {
                    "id": "456e7890-e89b-12d3-a456-426614174001",
                    "name": "Data Flow Diagram",
                    "type": "DFD-1.0.0"
                  }
                ]
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["Threat Model Sub-Resources"],
        "summary": "Create a new diagram",
        "description": "Creates a new diagram within the specified threat model",
        "operationId": "createThreatModelDiagram",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "threat_model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "maxLength": 36
            },
            "description": "Unique identifier of the threat model (UUID)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Diagram"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Diagram created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Diagram"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  }
}